last version

+++ src/classes/helperFunc/cleanup.cpp
@@ -7,8 +7,6 @@
 
 void cleanupCgi(FdManager &manager, std::map<int, int> &cgiToClient)
 {
-	if (manager.cgi_state == FINISHED)
-		return;
 
 	if (manager.stat_fd_to_cgi == NOT_FINISHED)
 	{
@@ -25,9 +23,12 @@ void cleanupCgi(FdManager &manager, std::map<int, int> &cgiToClient)
 		cgiToClient.erase(manager.from_cgi_fd);
 		manager.stat_fd_from_cgi = FINISHED;
 	}
-
-	kill(manager.cgi_pid, SIGKILL);
-	waitpid(manager.cgi_pid, NULL, 0);
+	if (manager.cgi_pid > 0)
+	{
+		kill(manager.cgi_pid, SIGKILL);
+		waitpid(manager.cgi_pid, NULL, 0);
+		manager.cgi_pid = -1;
+	}
 
 	manager.cgi_state = FINISHED;
 	manager.to_cgi_fd = -1;
